requirements.txt
smbus
ephem
RPi.GPIO
sysv_ipc
